+++ extensions/chatgpt/CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [Feature] - 2024-04-30
 
 - Feature: add `Summarize` command to summarize website and YouTube video
+- Refactor: model dropdown save model switch `localstorage` to `cache`
 
 ## [Feature] - 2024-04-09